Finished poly / decay seg function
It seems that the decay segments match matlab - at least I
checked the first one and it seems to. However, there are not
as many as I think there should be so I need to check this
against matlab.

I now have the array with the decay segment information
passed back to the calling function and I am moving on to the
ML section.

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>