Added framework for SPLIT & return values
Changes some function definitions, etc, so that I
could return all required parameters to callers.

Added framework to divide up the wav if necessary -
at the moment the wav is not divided  but it
just requires a for loop to be uncommented and SPLIT
to be redefined to (e.g.) 20 from 1.

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>