SVD example file using SVDLIBC
Checking this in so I don't loose it.

Calcs pinv() for 250 Hz least squares RT fit for the
Bach wav that I use for testing the algorithm.

I think there is a bug or something in the library that
makes it impossible to use the matrix if the values in
the matrix are large.

Howver, I discovered that if I divide the 2nd column of
the matrix by the max value in the matrix, I can just
multiply the RT by that max value in the end and I still
get the correct RT.

HASHTAG WINNING

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>