Fixed up SVDLIBC stuff: RT calculated now
I had to hack the maths a bit by normalising the
matrix as mentioned in a previous commit, but Matlab
was OK with the hack.

Tidied up SVDLIB stuff. Fixed a matrix calculation bug.
Also fixed a bug so that standard deviation is not
calculated if there was only 1 RT calculated.

Value is close to what Matlab gives for 250 Hz, but I
still have the hacked hardcoded a, b and alpha params.

Time to test the full program and compare to Matlab.

Fingers crossed!!

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>