Applying a more reasonable maxeval to coarse srch
Coarse search in Matlab only has 100*numvar
iterations before exit. Tested my code, and with
maxeval set to 100 instead of 300 for the coarse search
I get the same result for the first 3 calls.

There is a very slight difference for the fine search
when I reduce maxeval so I will leave it as-is for now

I must use COBYLA I think - most flexible algorithm.

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>