Changes:
RTs are now written to a file as well as STDOUT.
If a segment contains < 36 decays it is discarded.
Doing the coarse DR estimate (implemented to try and
optimise the code) is now optional and can be switched
on by defining a #define. Another #define sets the
threshold for the number of decay rates that are
acceptable in a segment; below this the segment is
discarded for obtaining RT estimates.

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>