Updating README & License files.
These files now explain that my code can include the
GCC DLLs and still be licensed under the MIT Licence.
This is because I provided the links to the source code
and agreed to provide access to it if the links go dead.
I am invoking the GCC Runtime Library Exception Clause.

Signed-off-by: Kim Jones <K.Jones20@edu.salford.ac.uk>